session validator now properly validates sessions
The session validator works as express middleware that sets
req.user to the structure of the logged in user.

I believe the best way to do this is with a cookie, rather than
send-the-session-in-the-request-body way the login function does
it. This will require refactoring the login function and probably
the frontend, but it's a better way.

Signed-off-by: Jude A Rich <judearich@protonmail.com>